agents/aegis-pm-fixed/requirements.txt
google-adk==1.27.2
google-genai==1.68.0
chromadb==1.1.1
redis==6.4.0
asyncpg==0.30.0
fastapi==0.135.1
httpx==0.28.1
structlog==25.4.0
networkx==3.4.2
pydantic==2.12.5
pydantic-settings==2.11.0
uvicorn==0.42.0
